# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cartes Scolaires",
        "description": "API complète pour la gestion des cartes scolaires",
        "contact": {
            "name": "Support API",
            "email": "support@cartes-scolaires.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api.cartes-scolaires.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/photos": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Liste des photos avec pagination",
                "operationId": "6976ed4acb075501fdc62ab3ae4f1538",
                "parameters": [
                    {
                        "name": "eleve_id",
                        "in": "query",
                        "description": "Filtrer par élève",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "brouillon",
                                "validee",
                                "refusee",
                                "archivee"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des photos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Uploader une photo d'élève",
                "operationId": "6afbf2a4293d76ace58310a9c4935330",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo",
                                    "eleve_id"
                                ],
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "eleve_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo uploadée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/{id}": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Détails d'une photo",
                "operationId": "1391ded23a5afa18bde665dea5be5df5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la photo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Photos"
                ],
                "summary": "Supprimer une photo",
                "operationId": "fc095f78b657d783539b4790853c8e11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo supprimée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/{id}/valider": {
            "patch": {
                "tags": [
                    "Photos"
                ],
                "summary": "Valider une photo",
                "operationId": "b95218ec65da24249e288491b4061b67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo validée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/{id}/refuser": {
            "patch": {
                "tags": [
                    "Photos"
                ],
                "summary": "Refuser une photo",
                "operationId": "0ee9f9e1552c1f02a87d640733dca98f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motif"
                                ],
                                "properties": {
                                    "motif": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo refusée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/statistiques": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Statistiques des photos",
                "operationId": "3f5aea89a7df0ec768c7fb6bbc303bcd",
                "responses": {
                    "200": {
                        "description": "Statistiques"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "bearerAuth",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}